+++ Runtime/UI/AutoGridLayoutGroup.cs
@@ -1,328 +1,196 @@
-using DataKeeper.Attributes;
 using UnityEngine;
 using UnityEngine.UI;
 
 namespace DataKeeper.UI
 {
     [AddComponentMenu("DataKeeper/UI/Auto Grid Layout Group")]
-    /// <summary>
-    /// Layout class to arrange child elements in a grid format.
-    /// </summary>
-    /// <remarks>
-    /// The GridLayoutGroup component is used to layout child layout elements in a uniform grid where all cells have the same size. The size and the spacing between cells is controlled by the GridLayoutGroup itself. The children have no influence on their sizes.
-    /// </remarks>
     public class AutoGridLayoutGroup : LayoutGroup
     {
-        /// <summary>
-        /// Which corner is the starting corner for the grid.
-        /// </summary>
         public enum Corner
         {
-            /// <summary>
-            /// Upper Left corner.
-            /// </summary>
             UpperLeft = 0,
-            /// <summary>
-            /// Upper Right corner.
-            /// </summary>
             UpperRight = 1,
-            /// <summary>
-            /// Lower Left corner.
-            /// </summary>
             LowerLeft = 2,
-            /// <summary>
-            /// Lower Right corner.
-            /// </summary>
             LowerRight = 3
         }
 
-        /// <summary>
-        /// The grid axis we are looking at.
-        /// </summary>
-        /// <remarks>
-        /// As the storage is a [][] we make access easier by passing a axis.
-        /// </remarks>
         public enum Axis
         {
-            /// <summary>
-            /// Horizontal axis
-            /// </summary>
             Horizontal = 0,
-            /// <summary>
-            /// Vertical axis.
-            /// </summary>
             Vertical = 1
         }
 
-        /// <summary>
-        /// Constraint type on either the number of columns or rows.
-        /// </summary>
         public enum Constraint
         {
-            /// <summary>
-            /// Constrain the number of columns to a specified number.
-            /// </summary>
             FixedColumnCount = 0,
-            /// <summary>
-            /// Constraint the number of rows to a specified number.
-            /// </summary>
             FixedRowCount = 1
         }
 
+        public enum CellSizeMode
+        {
+            AspectRatio = 0,
+            Fill = 1
+        }
+
+        private const float MinAspectRatio = 0.0001f;
+
         [SerializeField] protected Corner m_StartCorner = Corner.UpperLeft;
-
-        /// <summary>
-        /// Which corner should the first cell be placed in?
-        /// </summary>
-        public Corner startCorner { get { return m_StartCorner; } set { SetProperty(ref m_StartCorner, value); } }
-
         [SerializeField] protected Axis m_StartAxis = Axis.Horizontal;
-
-        /// <summary>
-        /// Which axis should cells be placed along first
-        /// </summary>
-        /// <remarks>
-        /// When startAxis is set to horizontal, an entire row will be filled out before proceeding to the next row. When set to vertical, an entire column will be filled out before proceeding to the next column.
-        /// </remarks>
-        public Axis startAxis { get { return m_StartAxis; } set { SetProperty(ref m_StartAxis, value); } }
-
-        [SerializeField, ReadOnlyInspector] protected Vector2 m_CellSize = new Vector2(100, 100);
-
+        [SerializeField] protected Constraint m_Constraint = Constraint.FixedColumnCount;
+        [SerializeField, Min(1)] protected int m_ConstraintCount = 2;
+        [SerializeField] protected CellSizeMode m_CellSizeMode = CellSizeMode.AspectRatio;
+        [SerializeField, Min(MinAspectRatio)] protected float m_AspectRatio = 1f;
         [SerializeField] protected Vector2 m_Spacing = Vector2.zero;
 
-        /// <summary>
-        /// The spacing to use between layout elements in the grid on both axises.
-        /// </summary>
-        public Vector2 spacing { get { return m_Spacing; } set { SetProperty(ref m_Spacing, value); } }
+        public Corner startCorner { get => m_StartCorner; set => SetProperty(ref m_StartCorner, value); }
+        public Axis startAxis { get => m_StartAxis; set => SetProperty(ref m_StartAxis, value); }
+        public Constraint constraint { get => m_Constraint; set => SetProperty(ref m_Constraint, value); }
+        public int constraintCount { get => m_ConstraintCount; set => SetProperty(ref m_ConstraintCount, Mathf.Max(1, value)); }
+        public CellSizeMode cellSizeMode { get => m_CellSizeMode; set => SetProperty(ref m_CellSizeMode, value); }
+        public float aspectRatio { get => m_AspectRatio; set => SetProperty(ref m_AspectRatio, Mathf.Max(MinAspectRatio, value)); }
+        public Vector2 spacing { get => m_Spacing; set => SetProperty(ref m_Spacing, value); }
 
-        [SerializeField] protected Constraint m_Constraint = Constraint.FixedColumnCount;
+        public Vector2 cellSize => m_CellSize;
+        public int columns => m_Columns;
+        public int rows => m_Rows;
 
-        /// <summary>
-        /// Which constraint to use for the GridLayoutGroup.
-        /// </summary>
-        /// <remarks>
-        /// Specifying a constraint can make the GridLayoutGroup work better in conjunction with a [[ContentSizeFitter]] component. When GridLayoutGroup is used on a RectTransform with a manually specified size, there's no need to specify a constraint.
-        /// </remarks>
-        public Constraint constraint { get { return m_Constraint; } set { SetProperty(ref m_Constraint, value); } }
+        private Vector2 m_CellSize;
+        private int m_Columns = 1;
+        private int m_Rows = 1;
 
-        
-        [SerializeField] protected int m_ConstraintCount = 2;
-
-        /// <summary>
-        /// How many cells there should be along the constrained axis.
-        /// </summary>
-        public int constraintCount { get { return m_ConstraintCount; } set { SetProperty(ref m_ConstraintCount, Mathf.Max(1, value)); } }
-
-        
-        [SerializeField] protected float m_AspectRatio = 1;
-        public float aspectRatio { get { return m_AspectRatio; } set { SetProperty(ref m_AspectRatio, Mathf.Max(0.001f, value)); } }
+        // The container size feeds the cell size on these axes, so it must not be reported back as a minimum.
+        public bool isWidthDriven => m_CellSizeMode == CellSizeMode.Fill || m_Constraint == Constraint.FixedColumnCount;
+        public bool isHeightDriven => m_CellSizeMode == CellSizeMode.Fill || m_Constraint == Constraint.FixedRowCount;
 
         protected AutoGridLayoutGroup()
-        {}
+        {
+        }
 
-        #if UNITY_EDITOR
+#if UNITY_EDITOR
         protected override void OnValidate()
         {
             base.OnValidate();
-            constraintCount = constraintCount;
-            aspectRatio = aspectRatio;
+            m_ConstraintCount = Mathf.Max(1, m_ConstraintCount);
+            m_AspectRatio = Mathf.Max(MinAspectRatio, m_AspectRatio);
         }
+#endif
 
-        #endif
-
-        /// <summary>
-        /// Called by the layout system to calculate the horizontal layout size.
-        /// Also see ILayoutElement
-        /// </summary>
         public override void CalculateLayoutInputHorizontal()
         {
             base.CalculateLayoutInputHorizontal();
 
-            int minColumns = 0;
-            int preferredColumns = 0;
-            if (m_Constraint == Constraint.FixedColumnCount)
-            {
-                minColumns = preferredColumns = m_ConstraintCount;
-                
-                m_CellSize.x = rectTransform.rect.width - (spacing.x * (constraintCount - 1) + padding.horizontal);
-                m_CellSize.x /= constraintCount;
-                m_CellSize.y = m_CellSize.x / aspectRatio;
-            }
-            else if (m_Constraint == Constraint.FixedRowCount)
-            {
-                minColumns = preferredColumns = Mathf.CeilToInt(rectChildren.Count / (float)m_ConstraintCount - 0.001f);
-                
-                m_CellSize.y = rectTransform.rect.height - (spacing.y * (constraintCount - 1) + padding.vertical);
-                m_CellSize.y /= constraintCount;
-                m_CellSize.x = m_CellSize.y * aspectRatio;
-            }
-            else
-            {
-                minColumns = 1;
-                preferredColumns = Mathf.CeilToInt(Mathf.Sqrt(rectChildren.Count));
-            }
-            
-            m_CellSize.x = Mathf.Max(0, m_CellSize.x);
-            m_CellSize.y = Mathf.Max(0, m_CellSize.y);
+            UpdateGrid();
+            UpdateCellSize();
 
-            SetLayoutInputForAxis(
-                padding.horizontal + (m_CellSize.x + spacing.x) * minColumns - spacing.x,
-                padding.horizontal + (m_CellSize.x + spacing.x) * preferredColumns - spacing.x,
-                -1, 0);
+            float width = padding.horizontal + m_CellSize.x * m_Columns + m_Spacing.x * (m_Columns - 1);
+            SetLayoutInputForAxis(isWidthDriven ? 0f : width, width, -1, 0);
         }
 
-        /// <summary>
-        /// Called by the layout system to calculate the vertical layout size.
-        /// Also see ILayoutElement
-        /// </summary>
         public override void CalculateLayoutInputVertical()
         {
-            int minRows = 0;
+            float height = padding.vertical + m_CellSize.y * m_Rows + m_Spacing.y * (m_Rows - 1);
+            SetLayoutInputForAxis(isHeightDriven ? 0f : height, height, -1, 1);
+        }
+
+        public override void SetLayoutHorizontal()
+        {
+            UpdateCellSize();
+            PlaceCells();
+        }
+
+        public override void SetLayoutVertical()
+        {
+            // Own height is final only now, and a height driven cell size also feeds the cell width,
+            // so both axes are placed again instead of only the vertical one.
+            UpdateCellSize();
+            PlaceCells();
+        }
+
+        private void UpdateGrid()
+        {
+            int count = rectChildren.Count;
+
             if (m_Constraint == Constraint.FixedColumnCount)
             {
-                minRows = Mathf.CeilToInt(rectChildren.Count / (float)m_ConstraintCount - 0.001f);
-            }
-            else if (m_Constraint == Constraint.FixedRowCount)
-            {
-                minRows = m_ConstraintCount;
+                m_Columns = Mathf.Max(1, m_ConstraintCount);
+                m_Rows = Mathf.Max(1, Mathf.CeilToInt(count / (float)m_Columns - 0.001f));
             }
             else
             {
-                float width = rectTransform.rect.width;
-                int cellCountX = Mathf.Max(1, Mathf.FloorToInt((width - padding.horizontal + spacing.x + 0.001f) / (m_CellSize.x + spacing.x)));
-                minRows = Mathf.CeilToInt(rectChildren.Count / (float)cellCountX);
+                m_Rows = Mathf.Max(1, m_ConstraintCount);
+                m_Columns = Mathf.Max(1, Mathf.CeilToInt(count / (float)m_Rows - 0.001f));
             }
-
-            float minSpace = padding.vertical + (m_CellSize.y + spacing.y) * minRows - spacing.y;
-            SetLayoutInputForAxis(minSpace, minSpace, -1, 1);
         }
 
-        /// <summary>
-        /// Called by the layout system
-        /// Also see ILayoutElement
-        /// </summary>
-        public override void SetLayoutHorizontal()
+        private void UpdateCellSize()
         {
-            SetCellsAlongAxis(0);
-        }
+            Rect rect = rectTransform.rect;
+            float cellWidth = Mathf.Max(0f, (rect.width - padding.horizontal - m_Spacing.x * (m_Columns - 1)) / m_Columns);
+            float cellHeight = Mathf.Max(0f, (rect.height - padding.vertical - m_Spacing.y * (m_Rows - 1)) / m_Rows);
 
-        /// <summary>
-        /// Called by the layout system
-        /// Also see ILayoutElement
-        /// </summary>
-        public override void SetLayoutVertical()
-        {
-            SetCellsAlongAxis(1);
-        }
-
-        private void SetCellsAlongAxis(int axis)
-        {
-            // Normally a Layout Controller should only set horizontal values when invoked for the horizontal axis
-            // and only vertical values when invoked for the vertical axis.
-            // However, in this case we set both the horizontal and vertical position when invoked for the vertical axis.
-            // Since we only set the horizontal position and not the size, it shouldn't affect children's layout,
-            // and thus shouldn't break the rule that all horizontal layout must be calculated before all vertical layout.
-            var rectChildrenCount = rectChildren.Count;
-            if (axis == 0)
+            if (m_CellSizeMode == CellSizeMode.Fill)
             {
-                // Only set the sizes when invoked for horizontal axis, not the positions.
-
-                for (int i = 0; i < rectChildrenCount; i++)
-                {
-                    RectTransform rect = rectChildren[i];
-
-                    m_Tracker.Add(this, rect,
-                        DrivenTransformProperties.Anchors |
-                        DrivenTransformProperties.AnchoredPosition |
-                        DrivenTransformProperties.SizeDelta);
-
-                    rect.anchorMin = Vector2.up;
-                    rect.anchorMax = Vector2.up;
-                    rect.sizeDelta = m_CellSize;
-                }
+                m_CellSize.x = cellWidth;
+                m_CellSize.y = cellHeight;
                 return;
             }
 
-            float width = rectTransform.rect.size.x;
-            float height = rectTransform.rect.size.y;
+            float aspect = Mathf.Max(MinAspectRatio, m_AspectRatio);
 
-            int cellCountX = 1;
-            int cellCountY = 1;
             if (m_Constraint == Constraint.FixedColumnCount)
             {
-                cellCountX = m_ConstraintCount;
-
-                if (rectChildrenCount > cellCountX)
-                    cellCountY = rectChildrenCount / cellCountX + (rectChildrenCount % cellCountX > 0 ? 1 : 0);
-            }
-            else if (m_Constraint == Constraint.FixedRowCount)
-            {
-                cellCountY = m_ConstraintCount;
-
-                if (rectChildrenCount > cellCountY)
-                    cellCountX = rectChildrenCount / cellCountY + (rectChildrenCount % cellCountY > 0 ? 1 : 0);
+                m_CellSize.x = cellWidth;
+                m_CellSize.y = cellWidth / aspect;
             }
             else
             {
-                if (m_CellSize.x + spacing.x <= 0)
-                    cellCountX = int.MaxValue;
-                else
-                    cellCountX = Mathf.Max(1, Mathf.FloorToInt((width - padding.horizontal + spacing.x + 0.001f) / (m_CellSize.x + spacing.x)));
-
-                if (m_CellSize.y + spacing.y <= 0)
-                    cellCountY = int.MaxValue;
-                else
-                    cellCountY = Mathf.Max(1, Mathf.FloorToInt((height - padding.vertical + spacing.y + 0.001f) / (m_CellSize.y + spacing.y)));
+                m_CellSize.y = cellHeight;
+                m_CellSize.x = cellHeight * aspect;
             }
+        }
 
-            int cornerX = (int)startCorner % 2;
-            int cornerY = (int)startCorner / 2;
+        private void PlaceCells()
+        {
+            int count = rectChildren.Count;
+            if (count == 0) return;
 
-            int cellsPerMainAxis, actualCellCountX, actualCellCountY;
-            if (startAxis == Axis.Horizontal)
+            int cellsPerMainAxis;
+            int actualColumns;
+            int actualRows;
+
+            if (m_StartAxis == Axis.Horizontal)
             {
-                cellsPerMainAxis = cellCountX;
-                actualCellCountX = Mathf.Clamp(cellCountX, 1, rectChildrenCount);
-                actualCellCountY = Mathf.Clamp(cellCountY, 1, Mathf.CeilToInt(rectChildrenCount / (float)cellsPerMainAxis));
+                cellsPerMainAxis = m_Columns;
+                actualColumns = Mathf.Clamp(m_Columns, 1, count);
+                actualRows = Mathf.Clamp(m_Rows, 1, Mathf.CeilToInt(count / (float)cellsPerMainAxis));
             }
             else
             {
-                cellsPerMainAxis = cellCountY;
-                actualCellCountY = Mathf.Clamp(cellCountY, 1, rectChildrenCount);
-                actualCellCountX = Mathf.Clamp(cellCountX, 1, Mathf.CeilToInt(rectChildrenCount / (float)cellsPerMainAxis));
+                cellsPerMainAxis = m_Rows;
+                actualRows = Mathf.Clamp(m_Rows, 1, count);
+                actualColumns = Mathf.Clamp(m_Columns, 1, Mathf.CeilToInt(count / (float)cellsPerMainAxis));
             }
 
-            Vector2 requiredSpace = new Vector2(
-                actualCellCountX * m_CellSize.x + (actualCellCountX - 1) * spacing.x,
-                actualCellCountY * m_CellSize.y + (actualCellCountY - 1) * spacing.y
-            );
-            Vector2 startOffset = new Vector2(
-                GetStartOffset(0, requiredSpace.x),
-                GetStartOffset(1, requiredSpace.y)
-            );
+            float requiredWidth = actualColumns * m_CellSize.x + (actualColumns - 1) * m_Spacing.x;
+            float requiredHeight = actualRows * m_CellSize.y + (actualRows - 1) * m_Spacing.y;
 
-            for (int i = 0; i < rectChildrenCount; i++)
+            float startX = GetStartOffset(0, requiredWidth);
+            float startY = GetStartOffset(1, requiredHeight);
+
+            int cornerX = (int)m_StartCorner % 2;
+            int cornerY = (int)m_StartCorner / 2;
+
+            for (int i = 0; i < count; i++)
             {
-                int positionX;
-                int positionY;
-                if (startAxis == Axis.Horizontal)
-                {
-                    positionX = i % cellsPerMainAxis;
-                    positionY = i / cellsPerMainAxis;
-                }
-                else
-                {
-                    positionX = i / cellsPerMainAxis;
-                    positionY = i % cellsPerMainAxis;
-                }
+                int columnIndex = m_StartAxis == Axis.Horizontal ? i % cellsPerMainAxis : i / cellsPerMainAxis;
+                int rowIndex = m_StartAxis == Axis.Horizontal ? i / cellsPerMainAxis : i % cellsPerMainAxis;
 
-                if (cornerX == 1)
-                    positionX = actualCellCountX - 1 - positionX;
-                if (cornerY == 1)
-                    positionY = actualCellCountY - 1 - positionY;
+                if (cornerX == 1) columnIndex = actualColumns - 1 - columnIndex;
+                if (cornerY == 1) rowIndex = actualRows - 1 - rowIndex;
 
-                SetChildAlongAxis(rectChildren[i], 0, startOffset.x + (m_CellSize[0] + spacing[0]) * positionX, m_CellSize[0]);
-                SetChildAlongAxis(rectChildren[i], 1, startOffset.y + (m_CellSize[1] + spacing[1]) * positionY, m_CellSize[1]);
+                RectTransform child = rectChildren[i];
+                SetChildAlongAxis(child, 0, startX + (m_CellSize.x + m_Spacing.x) * columnIndex, m_CellSize.x);
+                SetChildAlongAxis(child, 1, startY + (m_CellSize.y + m_Spacing.y) * rowIndex, m_CellSize.y);
             }
         }
     }